.emacs.d: rolled back to hjkl from jkil movement bindings in meow
Found that switching between Vim and Emacs is a lot harder with the “ergonomic”
movement bindings and since they are non-standard, the mental load is just not
worth it for me, since I am already accustomed to Vim bindings.

Also changed beginning of line to 's', insert to 'i' and append to 'a', for a
smoother transition between Vim and Emacs. I do not want to make all bindings
like Vim, since I don’t like w/b and $/^ at all and rather have them at u/o and
e/s.

I also changed z/Z to y/Y for undo/redo, since I found that it was too annoying
to hit, especially with the shift key, and I would often accidentally type it.
Also, y is more close to u, so it is more similar to Vim’s binding. It is also
too far away to accidentally hit it.