.emacs.d: Janet nrepl, ts symex replacements, janet-ts-mode fixes
Refactored symex replacements to be closer to the source, including tree-sitter
compatible functions, which do not work yet because of conflicts with built-in
tree-sitter (I think). I wanted to do that for easier merging of changes in the
repo later on.

Installed a Janet Netrepl client that uses Spork (a Janet lib). Not sure if I
need it yet, but it works.

Fixed some lines where janet-mode has not been replaced with janet-ts-mode.