.emacs.d: more meow prefix bindings, eldoc experiments, use-package
Disabled `use-package-always-ensure` since I want to manually declare what needs
to be installed. May change it back at some point.

Tried to upgrade eldoc manually, which worked, but Emacs will still use the
built-in version and there is no way to override it. My motivation was a newer
version of eldoc that enables `eldoc-echo-area-prefer-doc-buffer` (see
https://github.com/joaotavora/eglot/discussions/691#discussioncomment-719084).

I am quite dissatisfied with the eglot integration of eldoc, which is too noisy
for my taste. I wish it would work more like with Cider, where there is just
one line in the echo area for function signature and you can pop up the full
docs manually via a shortcut. It is frustrating to customize, but maybe things
will get better.