.emacs.d: commands to resize frame, some org & meow configuration
For some reason, Calc mode fails to sync the trail cursor when the window height
results in a chopped-off line at the bottom of the trail window. For this reason
I wrote some commands to set the frame height to a specific size (which is
different for different monitors), because this issue could not be fixed by just
setting `calc-window-height` to a different value. The size has to be readjusted
every time I change line height, font size or even the font itself. This is
rather annoying, but I cannot find another way around it at this time.