Emacs vterm meow-integration & shell-side config, sgrep
Removed clj-refactor due to connecting issues with CIDER to Clerks new
render-REPL.

Created a custom org-table converter for semicolon-separated tables.

Added org-mode hook to toggle org-appear when entering/exiting insert-mode in
meow, which is much less annoying than having symbols appear everytime the
cursor is above the characters.

Added integration with roswell (Common Lisp).

Experimented with Prots commands in popup frames with emacsclient to make a
command that opens Emacs Calc in a new frame with a system shortcut (via
Keyboard Maestro). It does work well but I’m not sure if I use it that often.

Added sgrep to shell config, which is a semantic grep tool useful for searching
with synonyms and inflections in text files.