emacs: various small things, gnu-apl-mode
Fixed macro recording in meow when not in meow-grab selection.

Deactivated yasnippet for now since I am unsure if I need it.

Installed gnu-apl-mode. It’s not as good as Dyalog, but FOS and works well from
Emacs.

Tried to write an artist-mode hook that disables non-essential minor modes
temporarily when using it, but the approach doesn’t seem to work.